ExerciseObject 만들어서 따로 관리하도록 수정
https://github.com/zimkjh/muscleKing/issues/33
