+++ Assets/GoogleMobileAds/Common/DummyClient.cs
@@ -26,7 +26,7 @@ namespace GoogleMobileAds.Common
     {
         public DummyClient()
         {
-            Debug.Log("Dummy " + MethodBase.GetCurrentMethod().Name);
+            // Debug.Log("Dummy " + MethodBase.GetCurrentMethod().Name);
         }
 
         // Disable warnings for unused dummy ad events.
@@ -71,7 +71,7 @@ namespace GoogleMobileAds.Common
 
         public void Initialize(string appId)
         {
-            Debug.Log("Dummy " + MethodBase.GetCurrentMethod().Name);
+            // Debug.Log("Dummy " + MethodBase.GetCurrentMethod().Name);
         }
 
         public void Initialize(Action<IInitializationStatusClient> initCompleteAction)
@@ -137,7 +137,7 @@ namespace GoogleMobileAds.Common
 
         public void LoadAd(AdRequest request)
         {
-            Debug.Log("Dummy " + MethodBase.GetCurrentMethod().Name);
+            // Debug.Log("Dummy " + MethodBase.GetCurrentMethod().Name);
         }
 
         public void ShowBannerView()
@@ -179,7 +179,7 @@ namespace GoogleMobileAds.Common
 
         public void CreateInterstitialAd(string adUnitId)
         {
-            Debug.Log("Dummy " + MethodBase.GetCurrentMethod().Name);
+            // Debug.Log("Dummy " + MethodBase.GetCurrentMethod().Name);
         }
 
         public bool IsLoaded()
