+++ Assets/Scripts/BigIntegerController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7ce4ff3e99044428b96aaafd8e5b4c7c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
