+++ Assets/Scripts/Objects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5d283716dcd1249ce81a1a3ae848c8f3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
