+++ Assets/Scripts/Objects/ExerciseObject.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c13b87b2e77094198940e355b685bd2f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
