+++ Assets/Scripts/Test.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ce37cfcf2e3ff4bc8bdd2df4533973c4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
